:boom: Created function: new_hybrid_interpolator
This is a true masterpiece, yet to be refined. This allows creating a linear combination of different interpolation methods to enforce ideal smoothness. On the other hand interpolating ISIG values seems like a bad idea. Only the original will be kept to have a track of which were the original entries.