:zap: Radically improved performance.
Setting all NaNs to pandas._libs.tslibs.nattype.NaT in columns that were supposed to be timedelta64[ns] is at least three orders of magnitude faster than applying pd.to_timedelta to the columns. I think this behaviour is rather odd. This is however optimal. Columns of this type are now parsed as object instead of StringType for this allows casting without errors.