:heavy_plus_sign: Added vaex.
This seems to be a promising technology to use, as the size of my data is ever-growing. Furthermore the desire to integrate everything into a single csv file (or database) requires the use of this technology. A year worth of 640G data seems to take up about 1 Go (giga octet) on disk as a csv file (4867773 rows, 52 columns). Even specifying dtypes on import, the resulting pandas.DataFrame is about 1.5 Go. This could pose problems in the future.